chore: break down attachments-part
into: attachment.tsx
      attachment-utils.ts
      attachments-block.tsx

update import and usages
